+++ change_log.md
@@ -11,6 +11,14 @@
 ## develop history ##
 ---
 
+
+- 2016/11/01
+    > 1 session.go:Session{maxPkgLen(int)} -> Session{maxPkgLen(int32)}
+    >
+    > 2 add remarks in session.go
+    >
+    > 2 version: 0.4.07
+
 - 2016/10/21
     > 1 session.go:(Session)RunEventLoop -> session.go:(Session)run
     >